requirements/base.txt
appdirs==1.4.3
dj-database-url==0.4.2
Django==1.11.1
django-allauth==0.31.0
django-dajax==0.9.2
django-dajaxice==0.7
django-jquery==3.1.0
django-json==0.14
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
functools32==3.2.3.post2
jsonschema==2.6.0
oauthlib==2.0.1
packaging==16.8
psycopg2==2.7.1
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
six==1.10.0
